requirements.txt
gradio
google-cloud-vision